found bug in double writing

+++ GdsSharp.Lib/Parsing/IGdsRecord.cs
@@ -3,5 +3,5 @@
 public interface IGdsRecord
 {
     ushort Code { get; }
-    int GetLength();
+    ushort GetLength();
 }
